requirements.txt
altair==5.5.0
appdirs==1.4.4
attrs==25.1.0
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
Flask==3.1.0
fonttools==4.55.5
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
narwhals==1.25.1
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pdfkit==1.0.0
pillow==11.1.0
plotly==6.0.0
pluggy==1.5.0
protobuf==5.29.3
pyarrow==19.0.0
pydeck==0.9.1
pyee==11.1.1
Pygments==2.19.1
pymongo==4.10.1
pyparsing==3.2.1
pyppeteer==2.0.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
streamlit==1.42.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==1.26.20
watchdog==6.0.0
websockets==10.4
Werkzeug==3.1.3
XlsxWriter==3.2.2
zipp==3.21.0
